Renamed cdbot folder
