+++ cdbot/cogs/general.py
