+++ cdbot/constants.py
