+++ cdbot/log.py
@@ -5,7 +5,7 @@ An implementation of a logging.Handler for sending messages to Discord
 import datetime
 import logging
 
-from cyberdisc_bot.constants import LOGGING_CHANNEL_ID
+from cdbot.constants import LOGGING_CHANNEL_ID
 from discord import Color, Embed
 from discord.ext import commands
