+++ cdbot/resources/AgentJSaysBlank.png
