+++ cdbot/resources/AngryJSaysBlank.png
