+++ cdbot/resources/JibhatSaysBlank.png
