+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "cyberdisc-bot"
+name = "cdbot"
 version = "0.1.0"
 description = "A Discord bot for the Cyber Discovery community."
 authors = ["Cyber Discovery Community <triggerhappymods@cyberdiscoverycommunity.uk>"]
@@ -26,7 +26,7 @@ safety = "=1.8.4"
 dodgy = "=0.1.9"
 
 [tool.poetry.scripts]
-cyberdisc-bot = "cyberdisc_bot:main"
+cdbot = "cdbot:main"
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
